Moved SUBSCRIPTION_QUERY to dependencies
